% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcp.R
\name{diffcp}
\alias{diffcp}
\title{diffcp function}
\usage{
diffcp(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object. Must contain assays GISTIC or a matrix of copy number values, assay should contain "GISTIC" in the name,
and Cohort (from `rnasubset` or `mutsubset`)}
}
\value{
data frame with frequency of mutation in high and low group with p-value from Chi-squared test
}
\description{
Computes gene with differential copy number between two subgroups selected by `rnasubset` or `mutsubset` by Chi-squared test
}
\examples{
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_ThresholdedByGene", "RPPAArray"), FALSE)
genome(lusc[[2]]) <- vapply(genome(lusc[[2]]), translateBuild, character(1L))
seqlevelsStyle(lusc[[2]]) <- "NCBI"
lusc2 <- qreduceTCGA(lusc)

#remove normal samples and just focus on primaries
tnmae <- splitAssays(lusc2, c("01", "11"))

lusc.t <- tnmae[, , grep("^01", names(tnmae))]
#split tumor and normal samples
lusc_tn <- splitAssays(lusc2, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_dcp <- diffcp(lusc_egfr)

}
