% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmethyl.R
\name{diffmethyl}
\alias{diffmethyl}
\title{diffmethyl function}
\usage{
diffmethyl(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object containing Methylation assay and Cohort assay from rnasubset}
}
\value{
data frame of significant differentially methylated islands between the two groups defined in rnasubset or mutsubset
}
\description{
Uses limma to find differentially methylated islands between the high and low groups
}
\examples{
#using data from the cureatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_ThresholdedByGene", "RPPAArray", "Methylation_methyl450"), FALSE)
#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
#subset LUSC data by expression of EGFR
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
#prepare methylation data for differential methylation
lusc_t.egfr2 <- methylprep(lusc_t.egfr)
#find differential methlyated islands between the high and low groups
egfr_diffmeth <- diffmethyl(lusc_t.egfr2)

}
