% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmut.R
\name{diffmut}
\alias{diffmut}
\title{diffmut function}
\usage{
diffmut(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object. Must contain assays Mutation (from curatedTCGA) and ExpressionLevel (from function rnasubset())}
}
\value{
data frame with frequency of mutation in high and low group with p-value from Chi-squared test
}
\description{
Computes differentially mutated genes between high and low group by Chi-squared test
}
\examples{
mae <- curatedTCGAData("SKCM", c("RNASeq2GeneNorm", "Mutation", "GISTIC", "RPPAArray") , FALSE)
gene <- "SOX10"
sox10 <- rnasubset(mae, gene, 10)
diffmut(sox10)

}
