% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffrna.R
\name{diffrna}
\alias{diffrna}
\title{diffrna function}
\usage{
diffrna(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object containing RNAseq assay and Cohort assay from rnasubset}
}
\value{
data frame of significant differentially expressed genes between the two groups defined in rnasubset or mutsubset
}
\description{
Uses voom to find differentially expressed genes between the high and low groups
}
\examples{
#using data from the cureatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_ThresholdedByGene", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_deg <- diffrna(lusc_t.egfr)

}
