% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffrppa.R
\name{diffrppa}
\alias{diffrppa}
\title{diffrppa function}
\usage{
diffrppa(mae, lfc = 1, pvalue = 0.05)
}
\arguments{
\item{mae}{MultiAssayExperiment object containing RNAseq assay and Cohort assay from rnasubset}

\item{lfc}{log fold change cutoff for differential protein expression, default = 1}

\item{pvalue}{pvalue cutoff for differential protein expression, default = 0.05}
}
\value{
data frame of significant differentially expressed proteins between the two groups defined in rnasubset or mutsubset
}
\description{
Uses limma to find differentially expressed proteins between the high and low groups
}
\examples{
#using data from the curatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_diffrppa <- diffrppa(lusc_t.egfr, 1, 0.05)

}
