% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffsurv.R
\name{diffsurv}
\alias{diffsurv}
\title{diffsurv function}
\usage{
diffsurv(mae, gene)
}
\arguments{
\item{mae}{MultiAssayExperiment object with Cohort assay generated from rnasubset or mutsubset.}

\item{gene}{character(1) gene of interest}
}
\value{
data frame of survival statistics
}
\description{
Returns a data table of median survival time and 95 confidence intervals of the two cohorts
}
\examples{
library(curatedTCGAData)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc2, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
diffsurv(lusc_t.egfr, "EGFR")

}
