% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genecoxsurv.R
\name{genecoxsurv}
\alias{genecoxsurv}
\title{genecoxsurv function}
\usage{
genecoxsurv(pat2, gene)
}
\arguments{
\item{pat2}{data frame generated by rnasubset function}

\item{gene}{character(1) gene of interest}
}
\value{
survminer styled Survival plot
}
\description{
performs cox univariate regression on the gene of interest
}
\examples{
data(skcm)
gene <- "SOX10"
sox10 <- rnasubset(pat, rna, gene, 10)
genecoxsurv(sox10, gene)

}
