% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchartgenelevel.R
\name{hchartgenelevel}
\alias{hchartgenelevel}
\title{hchartgenelevel function}
\usage{
hchartgenelevel(data)
}
\arguments{
\item{data}{data frame generated by patsubset function}
}
\value{
interactive scatter plot for gene level across all patients
}
\description{
Draws highchart scatter plot for gene levels
}
\examples{
gene <- "SOX10"
sox10 <- patsubset(pat, rna, gene, 10)
hchartgenelevel(sox10)

}

