% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello function}
\usage{
hello(who)
}
\arguments{
\item{who}{character(1) holds name of person}
}
\value{
character(1) displays name and number of characters
}
\description{
Outputs your name and counts characters
}
\examples{
hello("Martin")

}

