% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylprep.R
\name{methylprep}
\alias{methylprep}
\title{methylprep function}
\usage{
methylprep(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object containing Methylation assay and Cohort assay from rnasubset}
}
\value{
Returns the MultiAssayExperiment object with a new assay called SimpleMethyl contains a matrix of gene associated CpG islands and samples
}
\description{
Converts the methylation 450 SummarizedExperiment into a matrix of gene associated CpG Islands and samples
Uses CpG island definitions from `r Biocpkg(IlluminaHumanMethylation450kanno.ilmn12.hg19)`
}
\examples{
#using data from the cureatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_AllbyGene", "RPPAArray", "Methylation_methyl450"), FALSE)
#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
#prepare methylation data for simple differential methylation analysis
lusc_t <- methylprep(lusc_t)
#subset samples by top and bottom 10\% of EGFR expression
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
#look for differentially methylated islands
egfr_meth_deg <- diffmeth(lusc_t.egfr)

}
