% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutsubset.R
\name{mutsubset}
\alias{mutsubset}
\title{mutsubset function}
\usage{
mutsubset(mae, gene)
}
\arguments{
\item{mae}{MultiAssayExperiment object containing a simplifed Mutation assay generated from `qreduceTCGA`, must have assay with "Mutation" in the name}

\item{gene}{character(1) A gene to subset the mutation data}
}
\value{
Returns a MultiAssayExperiment object with an assay called Cohort that labels the samples according to "mutated" and "wild-type" for the gene of interest
}
\description{
Adds an assay called Cohort to a MultiAssayExperiment object to indicate which samples are mutated or wild-type for a gene of interest.
}
\examples{
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_ThresholdedByGene", "RPPAArray"), FALSE)
genome(lusc[[2]]) <- vapply(genome(lusc[[2]]), translateBuild, character(1L))
seqlevelsStyle(lusc[[2]]) <- "NCBI"
lusc2 <- qreduceTCGA(lusc)
#split tumor and normal samples
lusc_tn <- splitAssays(lusc2, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_egfr <- mutsubset(lusc_t, "EGFR")

}
