% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patsubset.R
\name{patsubset}
\alias{patsubset}
\title{patsubset function}
\usage{
patsubset(pat, rna, gene, percent)
}
\arguments{
\item{pat}{data frame with patient data, each patient is one row}

\item{rna}{data frame with RNAseq count values, genes in rows, patients in columns}

\item{gene}{character(1) Gene symbol}

\item{percent}{numeric(1) percentile of patients to compare 1-50}
}
\value{
Returns a data frame with additional columns for gene of interest expression levels and identifier for high, middle and low groups
}
\description{
Adds gene expression column and gene level columns to patient table
}
\examples{

patsubset(SKCMpat, SKCMrna, "SOX10", 10)

}

