% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotclin.R
\name{plotclin}
\alias{plotclin}
\title{plotclin function}
\usage{
plotclin(mae, var)
}
\arguments{
\item{mae}{MultiAssayExperiment object containing colData and Cohort assay from rnasubset}

\item{var}{vector of columns from colData for summarize, default is years_to_birth, gender, histological_type and pathologic stage}
}
\value{
gtsummary table of clinical features
}
\description{
Plots box or bar plots of clinical features of the two cohorts
}
\examples{
#using data from the cureatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_ThresholdedByGene", "RPPAArray", "Methylation_methyl450"), FALSE)
#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
#subset LUSC data by expression of EGFR
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
#Plot clinical differences between the high and low group
plotclin(lusc_t.egfr, "pathologic_stage")
}
