% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgenelevel.R
\name{plotgenelevel}
\alias{plotgenelevel}
\title{plotgenelevel function}
\usage{
plotgenelevel(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment generated by rnasubset function, must have assay named Cohort}
}
\value{
interactive scatter plot for gene level across all patients
}
\description{
Plots bar plot for gene levels (log scale)
}
\examples{
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_AllbyGene", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_egfr <- rnasubset(lusc_t, "EGFR", 10)
plotgenelevel(lusc_egfr)

}
