% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlycp.R
\name{plotlycp}
\alias{plotlycp}
\title{plotlycp function}
\usage{
plotlycp(pat2, cp, gene, level)
}
\arguments{
\item{pat2}{data frame generated by patsubset function}

\item{cp}{data frame of mutation matrix}

\item{gene}{character(1) gene of interest}

\item{level}{choose "high" or "low" population}
}
\value{
interactive oncoprint-like plot for top 15 genes with most copy number aberations
}
\description{
Draws plotly plot of copy number matrix
}
\examples{
gene <- "SOX10"
sox10 <- patsubset(pat, rna, gene, 10)
plotlycp(sox10, cp, gene, "high")

}

