% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlygenelevel.R
\name{plotlygenelevel}
\alias{plotlygenelevel}
\title{plotlygenelevel function}
\usage{
plotlygenelevel(data)
}
\arguments{
\item{data}{data frame generated by patsubset function}
}
\value{
interactive scatter plot for gene level across all patients
}
\description{
Draws plotly bar plot for gene levels
}
\examples{
gene <- "SOX10"
sox10.pat <- patsubset(pat, rna, gene, 10)
hchartgenelevel(sox10.pat)

}

