% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlymut.R
\name{plotlymut}
\alias{plotlymut}
\title{plotlymut function}
\usage{
plotlymut(mae, genemut, gene, num_genes)
}
\arguments{
\item{mae}{MultiAssayExperiment object. Must contain assays Mutation (from curatedTCGA) and ExpressionLevel (from function rnasubset())}

\item{genemut}{data frame generated from diffmut function}

\item{gene}{character(1) gene of interest}

\item{num_genes}{numeric(1) number of genes to plot}
}
\value{
interactive oncoprint-like plot for top differentially mutated genes
}
\description{
Draws plotly plot of mutations
}
\examples{
mae <- curatedTCGAData("SKCM", c("RNASeq2GeneNorm", "Mutation", "GISTIC", "RPPAArray") , FALSE)
gene <- "SOX10"
sox10.pat <- rnasubset(mae, gene, 10)
sox10.mut <- diffmut(sox10.pat)
plotlymut(sox10.pat, mut, sox10.mut, gene)

}
