% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmethylheat.R, R/plotmethylregion.R
\name{plotmethylheat}
\alias{plotmethylheat}
\title{plotmethylheat function}
\usage{
plotmethylheat(mae, dmi, gene, n = 100)

plotmethylheat(mae, dmi, gene, n = 100)
}
\arguments{
\item{mae}{MultiAssayExperiment object containing SimpleMethyl assay and Cohort assay from rnasubset}

\item{dmi}{data frame of differential methylated islands from diffmethyl function}

\item{gene}{character(1) Gene symbol of gene of interest}

\item{n}{numeric(1) number of genes to plot for the heatmap, default = 100}

\item{region}{character(1) region of interest}
}
\value{
ComplexHeatmap of top 100 significant differentially methylated islands

ComplexHeatmap of top 100 significant differentially methylated islands
}
\description{
Uses ComplexHeatmap to draw heatmap

Uses Gviz to plot out the methylation Beta values across the CpG island
}
\examples{
#using data from the curatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("RNASeq2GeneNorm", "Methylation_methyl450"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
#prepare methylation data for differential methylation
lusc_t.egfr2 <- methylprep(lusc_t.egfr)
#find differential methlyated islands between the high and low groups
egfr_diffmeth <- diffmethyl(lusc_t.egfr2)
plotmethylheat(lusc_t.egfr2, egfr_diffmeth, "EGFR", 100)

#using data from the curatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_ThresholdedByGene", "RPPAArray", "Methylation_methyl450"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
#prepare methylation data for differential methylation
lusc_t.egfr2 <- methylprep(lusc_t.egfr)
#find differential methlyated islands between the high and low groups
egfr_diffmeth <- diffmethyl(lusc_t.egfr2)
plotmethylheat(lusc_t.egfr2, egfr_diffmeth, "EGFR", 100)

}
