% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmut.R
\name{plotmut}
\alias{plotmut}
\title{plotmut function}
\usage{
plotmut(mae, genemut, gene, n = 20)
}
\arguments{
\item{mae}{MultiAssayExperiment object. Must contain assays simplified Mutation (generated by `qreduceTCGA`)
or a binary matrix of mutated genes by sample, assay should contain "Mutation" in the name,
and Cohort (from `rnasubset` or `mutsubset`)}

\item{genemut}{data frame generated from diffmut function}

\item{gene}{character(1) gene of interest}

\item{n}{numeric(1) number of genes to plot default = 20}
}
\value{
oncoprint-like plot for top differentially mutated genes
}
\description{
Draws plot of mutations
}
\examples{
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_AllbyGene", "RPPAArray"), FALSE)
genome(lusc[[2]]) <- vapply(genome(lusc[[2]]), translateBuild, character(1L))
seqlevelsStyle(lusc[[2]]) <- "NCBI"
lusc2 <- qreduceTCGA(lusc)

#remove normal samples and just focus on primaries
tnmae <- splitAssays(lusc2, c("01", "11"))

lusc.t <- tnmae[, , grep("^01", names(tnmae))]
#split tumor and normal samples
lusc_tn <- splitAssays(lusc2, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_dm <- diffmut(lusc_egfr)
plotmut(lusc_egfr, egfr_dm, "EGFR", 20)

}
