% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotreact.R
\name{plotreact}
\alias{plotreact}
\title{rnareactplot function}
\usage{
plotreact(react, deg, graph, n)
}
\arguments{
\item{react}{data frame generated by rnareact function}

\item{deg}{data frame generated by rnadeg function}

\item{graph}{character(1) type of graph to plot, (dot, map, cnet)}

\item{n}{numeric(1) number of pathways to plot}
}
\value{
dotplot or enrichment pathway or cnetmap of Reactome gene sets
}
\description{
Uses ReactomePA to find enriched pathways from the RNAseq differential gene analysis
}
\examples{
data(skcm)
gene <- "SOX10"
sox10.pat <- rnasubset(pat, rna, gene, 10)
sox10.deg <- rnadeg(sox10.pat, rna)
sox10.react <- rnareact(sox10.deg)
plotreact(sox10.react, sox10.deg, "dot", 15)
plotreact(sox10.react, sox10.deg, "map", 15)
plotreact(sox10.react, sox10.deg, "cnet", 5)

}
