% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrnaheat.R
\name{plotrnaheat}
\alias{plotrnaheat}
\title{plotrnaheat function}
\usage{
plotrnaheat(mae, deg, gene, n = 100)
}
\arguments{
\item{mae}{MultiAssayExperiment object containing RNAseq assay and Cohort assay from rnasubset}

\item{deg}{data frame of differential expressed genes from rnadeg function}

\item{gene}{character(1) Gene symbol of gene of interest}

\item{n}{numeric(1) number of genes to plot for the heatmap, default = 100}
}
\value{
ComplexHeatmap of top 100 significant differentially expressed genes
}
\description{
Uses ComplexHeatmap to draw heatmap
}
\examples{
#using data from the cureatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_ThresholdedByGene", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_deg <- diffrna(lusc_t.egfr)
plotrnaheat(lusc_t.egfr, egfr_deg, "EGFR", 100)

}
