% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrppaheat.R
\name{plotrppaheat}
\alias{plotrppaheat}
\title{plotrppaheat function}
\usage{
plotrppaheat(mae, dpe, gene, n)
}
\arguments{
\item{mae}{MultiAssayExperiment object. Must contain assays Mutation (from curatedTCGA) and Cohort (from function rnasubset())}

\item{gene}{character(1) Gene symbol of gene of interest}
}
\value{
ComplexHeatmap of top 100 significant differentially expressed genes
}
\description{
Uses ComplexHeatmap to draw heatmap for RPPA data
}
\examples{
#using data from the curatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_diffrppa <- diffrppa(lusc_t.egfr, 0.1, 0.4)
plotrppaheat(lusc_t.egfr, egfr_diffrppa, "EGFR", 50)

}
