% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsurv.R
\name{plotsurv}
\alias{plotsurv}
\title{plotsurv function}
\usage{
plotsurv(mae, gene)
}
\arguments{
\item{mae}{MultiAssayExperiment object with Cohort assay generated from rnasubset or mutsubset.}

\item{gene}{character(1) gene of interest}
}
\value{
survminer styled Survival plot
}
\description{
Draws survminer survival plot for high and low gene populations
}
\examples{
library(curatedTCGAData)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
plotsurv(lusc_t.egfr, "EGFR")

}
