% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnadeg.R
\name{rnadeg}
\alias{rnadeg}
\title{rnadeg function}
\usage{
rnadeg(pat2, rna)
}
\arguments{
\item{pat2}{data frame generated by patsubset function}

\item{rna}{data frame of RNAseq values}
}
\value{
data frame of RNAseq values for only patients in the high and low gene expression groups
}
\description{
Uses voom to find differentially expressed genes between the high and low groups
}
\examples{
gene <- "SOX10"
sox10.pat <- patsubset(pat, rna, gene, 10)
sox10.rna <- rnsubset(sox10.pat, rna)

}

