% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnagage.R
\name{rnagage}
\alias{rnagage}
\title{rnagage function}
\usage{
rnagage(deg, lookup)
}
\arguments{
\item{lookup}{data frame of entrez ID and gene name pairs}

\item{rnadeg}{data frame generated by topTable of limma output}
}
\value{
data frame of RNAseq values for only patients in the high and low gene expression groups
}
\description{
Uses gage to find enriched pathways from the RNAseq differential gene analysis
}
\examples{
gene <- "SOX10"
sox10.pat <- patsubset(pat, rna, gene, 10)
sox10.deg <- rnadeg(sox10.pat, rna)
sox10.gage <- rnagage(sox10.deg, lookup)

}

