% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnagsva.R
\name{rnagsva}
\alias{rnagsva}
\title{rnagsva function}
\usage{
rnagsva(pat2, rna)
}
\arguments{
\item{pat2}{data frame generated by patsubset function}

\item{rna}{data frame of RNAseq values}
}
\value{
data frame of gsva values for all samples
}
\description{
Uses GSVA to find gene set enrichment across samples
}
\examples{
gene <- "SOX10"
sox10.pat <- patsubset(pat, rna, gene, 10)
sox10.deg <- rnadeg(sox10.pat, rna)
sox10.gsva <- rnagsva(sox10.pat, rna)

}

