% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnagsvasig.R
\name{rnagsvasig}
\alias{rnagsvasig}
\title{rnagsvasig function}
\usage{
rnagsvasig(pat2, rnagsva_out)
}
\arguments{
\item{pat2}{data frame generated by rnasubset function}

\item{rnagsva}{data frame generated by rnagsva}
}
\value{
data frame of logFC values for gene sets
}
\description{
Uses GSVA to find gene set enrichment across samples
}
\examples{
data(skcm)
gene <- "SOX10"
sox10.pat <- rnasubset(pat, rna, gene, 10)
sox10.gsva <- rnagsva(sox10.pat, rna)
sox10.gsva.sig <- rnagsvasig(sox10.pat, sox10.gsva)

}
