% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaheat.R
\name{rnaheat}
\alias{rnaheat}
\title{rnaheat function}
\usage{
rnaheat(pat2, rna, deg, gene, n = NULL)
}
\arguments{
\item{pat2}{data frame generated by rnasubset function}

\item{rna}{data frame of RNAseq values}

\item{deg}{data frame of differential expressed genes from rnadeg function}

\item{gene}{character(1) Gene symbol of gene of interest}

\item{n}{numeric(1) Number of genes you want to plot, default is all significant genes}
}
\value{
ComplexHeatmap of top 100 significant differentially expressed genes
}
\description{
Uses ComplexHeatmap to draw heatmap
}
\examples{
data(skcm)
gene <- "SOX10"
sox10.pat <- rnasubset(pat, rna, gene, 50)
sox10.deg <- rnadeg(sox10.pat, rna)
rnaheat(sox10.pat, rna, sox10.deg, gene, n)

}
