% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnamethdeg.R
\name{rnamethdeg}
\alias{rnamethdeg}
\title{rnamethdeg function}
\usage{
rnamethdeg(pat2, meth)
}
\arguments{
\item{pat2}{data frame generated by rnasubset function}

\item{meth}{data frame of CpG Island methylation values, methyltest6}
}
\value{
data frame of significant differentially methylated islands between the two groups defined in rnasubset or mutsubset
}
\description{
Uses limma to find differentially methylated islands between the high and low groups
}
\examples{
data(skcm)
gene <- "SOX10"
sox10.rna <- rnasubset(pat, rna, gene, 10)
sox10.methdeg <- rnamethdeg(sox10.rna, meth)

}
