% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnareact.R, R/rnareactdot.R
\name{rnareact}
\alias{rnareact}
\title{rnareact function}
\usage{
rnareact(rnadeg, lookup, n, graph)

rnareact(rnadeg, lookup, n, graph)
}
\arguments{
\item{rnadeg}{data frame generated by topTable of limma output}

\item{lookup}{data frame of entrez ID and gene name pairs}

\item{n}{numeric(1) number of pathways to plot}

\item{graph}{character(1) type of graph to plot, (dot, map, cnet)}

\item{rnadeg}{data frame generated by topTable of limma output}

\item{lookup}{data frame of entrez ID and gene name pairs}

\item{n}{numeric(1) number of pathways to plot}

\item{graph}{character(1) type of graph to plot, (dot, map, cnet)}
}
\value{
dotplot or enrichment pathway or cnetmap of Reactome gene sets

dot plot of Reactome enrichment
}
\description{
Uses ReactomePA to find enriched pathways from the RNAseq differential gene analysis

Uses ReactomePA to find enriched pathways from the RNAseq differential gene analysis
}
\examples{
gene <- "SOX10"
sox10.pat <- patsubset(pat, rna, gene, 10)
sox10.deg <- rnadeg(sox10.pat, rna)
rnareact(sox10.deg, lookup, 10, "dot")


gene <- "SOX10"
sox10.pat <- patsubset(pat, rna, gene, 10)
sox10.deg <- rnadeg(sox10.pat, rna)
rnareactdot(sox10.deg, lookup, 20)

}

