% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnasubset.R
\name{rnasubset}
\alias{rnasubset}
\title{rnasubset function}
\usage{
rnasubset(pat2, rna)
}
\arguments{
\item{pat2}{data frame generated by patsubset function}

\item{rna}{data frame of RNAseq values}
}
\value{
data frame of RNAseq values for only patients in the high and low gene expression groups
}
\description{
subsets rnaseq table for patient chosen in analysis
}
\examples{
gene <- "SOX10"
sox10.pat <- patsubset(pat, rna, gene, 10)
sox10.rna <- rnsubset(sox10.pat, rna)

}

