% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffrppa.R, R/rppadeg.R
\name{rppadeg}
\alias{rppadeg}
\title{diffrppa function}
\usage{
rppadeg(mae)

rppadeg(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object containing RNAseq assay and Cohort assay from rnasubset}

\item{mae}{MultiAssayExperiment object containing RNAseq assay and Cohort assay from rnasubset}
}
\value{
data frame of significant differentially expressed proteins between the two groups defined in rnasubset or mutsubset

data frame of significant differentially expressed proteins between the two groups defined in rnasubset or mutsubset
}
\description{
Uses limma to find differentially expressed proteins between the high and low groups

Uses limma to find differentially expressed proteins between the high and low groups
}
\examples{
#using data from the cureatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_AllbyGene", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_diffrppa <- diffrppa(lusc_t.egfr)

#using data from the cureatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_AllbyGene", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_deg <- rppadeg(lusc_t.egfr)

}
