% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppaheat.R
\name{rppaheat}
\alias{rppaheat}
\title{rppaheat function}
\usage{
rppaheat(mae, gene)
}
\arguments{
\item{mae}{MultiAssayExperiment object. Must contain assays Mutation (from curatedTCGA) and ExpressionLevel (from function rnasubset())}

\item{gene}{character(1) Gene symbol of gene of interest}
}
\value{
ComplexHeatmap of top 100 significant differentially expressed genes
}
\description{
Uses ComplexHeatmap to draw heatmap for RPPA data
}
\examples{
mae <- curatedTCGAData("SKCM", c("RNASeq2GeneNorm", "Mutation", "GISTIC", "RPPAArray") , FALSE)
gene <- "SOX10"
sox10 <- rnasubset(mae, gene, 10)
rppaheat(sox10, gene)
}
