% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samnum.R
\name{samnum}
\alias{samnum}
\title{samnum function}
\usage{
samnum(data)
}
\arguments{
\item{data}{a multiassayexperiment object}
}
\value{
data frame with assay names and number of samples in each assay
}
\description{
Counts how many samples are in each assay
}
\examples{

samnum(data)

}

