% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgashiny.R
\name{tcgashiny}
\alias{tcgashiny}
\title{tcgashiny function}
\usage{
tcgashiny(pat, rna, mut, cp, rppa)
}
\arguments{
\item{pat}{data frame containing clinical data}

\item{rna}{data frame containing RNAseq counts}

\item{mut}{data frame containing binary matrix of mutated genes}

\item{cp}{data frame containing copy number values}

\item{rppa}{data frame containing RPPA values}
}
\value{
shiny app that explores the 5 datasets
}
\description{
Launches local Shiny app to explore your dataset
}
\examples{
data(skcm)
tcgashiny(pat, rna, mut, cp, rppa)

}
