% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcp.R
\name{diffcp}
\alias{diffcp}
\title{diffcp function}
\usage{
diffcp(pat2, cp)
}
\arguments{
\item{pat2}{data frame generated by patsubset function}

\item{cp}{data frame of copy number matrix}
}
\value{
data frame with frequency of copy number changes in high and low group with p-value from Chi-squared test
}
\description{
Computes differential copy number genes between high and low group by Chi-squared test
}
\examples{
gene <- "SOX10"
sox10.pat <- patsubset(pat, rna, gene, 10)
diffcp(sox10.pat, cp)

}

