% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmut.R
\name{diffmut}
\alias{diffmut}
\title{diffmut function}
\usage{
diffmut(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object. Must contain assays Mutation (from curatedTCGA) and ExpressionLevel (from function rnasubset())}
}
\value{
data frame with frequency of mutation in high and low group with p-value from Chi-squared test
}
\description{
Computes differentially mutated genes between high and low group by Chi-squared test
}
\examples{
library(curatedTCGAData)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTICT", "RPPAArray"), FALSE)
genome(lusc[[2]])
genome(lusc[[2]]) <- vapply(genome(lusc[[2]]), translateBuild, character(1L))

seqlevelsStyle(lusc[[2]]) <- "NCBI"
genome(lusc[[2]])

lusc2 <- qreduceTCGA(lusc)

#remove normal samples and just focus on primaries
tnmae <- splitAssays(lusc2, c("01", "11"))

lusc.t <- tnmae[, , grep("^01", names(tnmae))]
#split tumor and normal samples
lusc_tn <- splitAssays(lusc2, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]

}
