% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmut.R
\name{diffmut}
\alias{diffmut}
\title{diffmut function}
\usage{
diffmut(pat2, mut)
}
\arguments{
\item{pat2}{data frame generated by rnasubset function}

\item{mut}{data frame of mutation matrix}
}
\value{
data frame with frequency of mutation in high and low group with p-value from Chi-squared test
}
\description{
Computes differentially mutated genes between high and low group by Chi-squared test
}
\examples{
data(skcm)
gene <- "SOX10"
sox10 <- rnasubset(pat, rna, gene, 10)
diffmut(sox10, mut, gene)

}

