% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffmut.R
\name{diffmut}
\alias{diffmut}
\title{diffmut function}
\usage{
diffmut(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object. Must contain assays simplified Mutation (generated by `qreduceTCGA`)
or a binary matrix of mutated genes by sample, assay should contain "Mutation" in the name,
and Cohort (from `rnasubset` or `mutsubset`)}
}
\value{
data frame with frequency of mutation in high and low group with p-value from Chi-squared test
}
\description{
Computes differentially mutated genes between two subgroups selected by `rnasubset` or `mutsubset` by Chi-squared test
}
\examples{
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm"), FALSE)
mut_assay <- grep("Mutation", names(lusc))
genome(lusc[[mut_assay]]) <- vapply(genome(lusc[[mut_assay]]), translateBuild, character(1L))
seqlevelsStyle(lusc[[mut_assay]]) <- "NCBI"
lusc2 <- qreduceTCGA(lusc)

#remove normal samples and just focus on primaries
tnmae <- splitAssays(lusc2, c("01", "11"))

lusc.t <- tnmae[, , grep("^01", names(tnmae))]
#split tumor and normal samples
lusc_tn <- splitAssays(lusc2, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_dm <- diffmut(lusc_t.egfr)

}
