% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesurv.R, R/plotsurv.R
\name{genesurv}
\alias{genesurv}
\title{genesurv function}
\usage{
genesurv(mae, gene)

genesurv(mae, gene)
}
\arguments{
\item{mae}{MultiAssayExperiment object with Cohort assay generated from rnasubset or mutsubset.}

\item{gene}{character(1) gene of interest}

\item{mae}{MultiAssayExperiment object with Cohort assay generated from rnasubset or mutsubset.}

\item{gene}{character(1) gene of interest}
}
\value{
survminer styled Survival plot

survminer styled Survival plot
}
\description{
Draws survminer survival plot for high and low gene populations

Draws survminer survival plot for high and low gene populations
}
\examples{
library(curatedTCGAData)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_AllbyGene", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc2, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_egfr <- rnasubset(lusc_t, "EGFR", 10)
genesurv(lusc_egfr, "EGFR")

library(curatedTCGAData)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTIC_ThresholdedByGene", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc2, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_egfr <- rnasubset(lusc_t, "EGFR", 10)
plotsurv(lusc_egfr, "EGFR")

}
