% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesurv.R
\name{genesurv}
\alias{genesurv}
\title{genesurv function}
\usage{
genesurv(pat2, gene)
}
\arguments{
\item{pat2}{data frame generated by rnasubset function}

\item{gene}{character(1) gene of interest}
}
\value{
survminer styled Survival plot
}
\description{
Draws survminer survival plot for high and low gene populations
}
\examples{
data(skcm)
gene <- "SOX10"
sox10 <- rnasubset(pat, rna, gene, 10)
genesurv(sox10, gene)

}

