% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutsubset.R
\name{mutsubset}
\alias{mutsubset}
\title{mutsubset function}
\usage{
mutsubset(pat, rna, mut, gene)
}
\arguments{
\item{pat}{data frame with patient data, each patient is one row}

\item{rna}{data frame with RNAseq count values, genes in rows, patients in columns}

\item{mut}{data frame with binary mutation, one gene per row, one patient per column}

\item{gene}{character(1) Gene symbol}
}
\value{
Returns a data frame with mutation of gene as WT or mut and gene expression
}
\description{
Adds mutation column to patient table
}
\examples{
data(skcm)
mutsubset(pat, rna, mut, "SOX10")

}

