% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlycp.R
\name{plotlycp}
\alias{plotlycp}
\title{plotlycp function}
\usage{
plotlycp(mae, genecp, gene, n = 20)
}
\arguments{
\item{mae}{MultiAssayExperiment object. Must contain assays GISTIC or a matrix of copy number values, assay should contain "GISTIC" in the name,
and Cohort (from `rnasubset` or `mutsubset`)}

\item{genecp}{data frame generated by diffcp function}

\item{gene}{character(1) gene of interest}

\item{n}{numeric(1) number of genes to plot default is 20}
}
\value{
interactive oncoprint-like plot for top 20 differentially copy number altered genes
}
\description{
Draws plotly plot of copy number matrix
}
\examples{
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTICT", "RPPAArray"), FALSE)
genome(lusc[[2]]) <- vapply(genome(lusc[[2]]), translateBuild, character(1L))
seqlevelsStyle(lusc[[2]]) <- "NCBI"
lusc2 <- qreduceTCGA(lusc)

#remove normal samples and just focus on primaries
tnmae <- splitAssays(lusc2, c("01", "11"))

lusc.t <- tnmae[, , grep("^01", names(tnmae))]
#split tumor and normal samples
lusc_tn <- splitAssays(lusc2, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_dcp <- diffcp(lusc_egfr)
plotlycp(lusc_egfr, egfr_dcp, "EGFR", 20)

}
