% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlycp.R
\name{plotlycp}
\alias{plotlycp}
\title{plotlycp function}
\usage{
plotlycp(pat2, cp, genecp, gene)
}
\arguments{
\item{pat2}{data frame generated by rnasubset function}

\item{cp}{data frame of mutation matrix}

\item{genecp}{data frame generated by diffcp function}

\item{gene}{character(1) gene of interest}
}
\value{
interactive oncoprint-like plot for top 20 differentially copy number altered genes
}
\description{
Draws plotly plot of copy number matrix
}
\examples{
data(skcm)
gene <- "SOX10"
sox10.pat <- rnasubset(pat, rna, gene, 10)
sox10.cp <- diffcp(sox10.pat, cp)
plotlycp(sox10.pat, cp, sox10.cp, gene)

}
