% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlygenelevel.R
\name{plotlygenelevel}
\alias{plotlygenelevel}
\title{plotlygenelevel function}
\usage{
plotlygenelevel(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment generated by rnasubset function, must have assay named Cohort}
}
\value{
interactive scatter plot for gene level across all patients
}
\description{
Draws plotly bar plot for gene levels (log scale)
}
\examples{
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTICT", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_egfr <- rnasubset(lusc_t, "EGFR", 10)
plotlygenelevel(lusc_egfr)

}
