% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlymut.R
\name{plotlymut}
\alias{plotlymut}
\title{plotlymut function}
\usage{
plotlymut(pat2, mut, genemut, gene)
}
\arguments{
\item{pat2}{data frame generated by rnasubset function}

\item{mut}{data frame of mutation matrix}

\item{genemut}{data frame generated from diffmut function}

\item{gene}{character(1) gene of interest}
}
\value{
interactive oncoprint-like plot for top 20 differentially mutated genes
}
\description{
Draws plotly plot of mutations
}
\examples{
data(skcm)
gene <- "SOX10"
sox10.pat <- rnasubset(pat, rna, gene, 10)
sox10.mut <- diffmut(sox10.pat, mut)
plotlymut(sox10.pat, mut, sox10.mut, gene)

}
