% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlymut.R
\name{plotlymut}
\alias{plotlymut}
\title{plotlymut function}
\usage{
plotlymut(pat2, mut, gene, level)
}
\arguments{
\item{pat2}{data frame generated by patsubset function}

\item{mut}{data frame of mutation matrix}

\item{gene}{character(1) gene of interest}

\item{level}{choose "high" or "low" population}
}
\value{
interactive oncoprint-like plot for top 15 most mutated genes
}
\description{
Draws plotly plot of mutations
}
\examples{
gene <- "SOX10"
sox10 <- patsubset(pat, rna, gene, 10)
plotlymut(sox10, mut, gene, "high")

}

