% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotreact.R
\name{plotreact}
\alias{plotreact}
\title{rnareactplot function}
\usage{
plotreact(react, deg, graph, n = 10)
}
\arguments{
\item{react}{data frame generated by rnareact function}

\item{deg}{data frame generated by rnadeg function}

\item{graph}{character(1) type of graph to plot, (dot, map, cnet)}

\item{n}{numeric(1) number of pathways to plot, default is 10}
}
\value{
dotplot or enrichment pathway or cnetmap of Reactome gene sets
}
\description{
Uses ReactomePA to find enriched pathways from the RNAseq differential gene analysis
}
\examples{
#using data from the cureatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTICT", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_deg <- rnadeg(lusc_t.egfr)
egfr_react <- rnareact(egfr_deg)
plotreact(egfr_react, egfr_deg, "dot", 10)
plotreact(egfr_react, egfr_deg, "map", 10)
plotreact(egfr_react, egfr_deg, "cnet", 10)
}
