% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnagage.R
\name{rnagage}
\alias{rnagage}
\title{rnagage function}
\usage{
rnagage(deg)
}
\arguments{
\item{rnadeg}{data frame generated from rnadeg}
}
\value{
data frame of RNAseq values for only patients in the high and low gene expression groups
}
\description{
Uses gage to find enriched pathways from the RNAseq differential gene analysis
}
\examples{
#using data from the cureatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("Mutation", "RNASeq2GeneNorm", "GISTICT", "RPPAArray"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_deg <- rnadeg(lusc_t.egfr)
egfr_path <- rnagage(egfr_deg)

}
