% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnagsvaheat.R
\name{rnagsvaheat}
\alias{rnagsvaheat}
\title{rnagsvaheat function}
\usage{
rnagsvaheat(pat2, rna_gsva, rna_gsvasig, gene)
}
\arguments{
\item{pat2}{data frame generated by rnasubset function}

\item{gene}{character(1) Gene symbol of gene of interest}

\item{rnagsva}{data frame of GSVA values}

\item{rnasgsvasig}{data frame of significant gene sets from rnagsvasig function}
}
\value{
ComplexHeatmap of top 100 significant gene sets
}
\description{
Uses ComplexHeatmap to draw heatmap for GSVA analysis
}
\examples{
data(skcm)
gene <- "SOX10"
sox10.pat <- rnasubset(pat, rna, gene, 5)
sox10.gsva <- rnagsva(sox10.pat, rna)
sox10.gsva.sig <- rnagsvasig(sox10.pat, sox10.gsva)
rnagsvaheat(sox10.pat, sox10.gsva, sox10.gsva.sig, gene)

}
