% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaheat.R
\name{rnaheat}
\alias{rnaheat}
\title{rnaheat function}
\usage{
rnaheat(pat2, rna, deg, gene)
}
\arguments{
\item{pat2}{data frame generated by patsubset function}

\item{rna}{data frame of RNAseq values}

\item{deg}{data frame of differential expressed genes from rnadeg function}

\item{gene}{character(1) Gene symbol of gene of interest}
}
\value{
ComplexHeatmap of top 100 significant differentially expressed genes
}
\description{
Uses ComplexHeatmap to draw heatmap
}
\examples{
gene <- "SOX10"
sox10.pat <- patsubset(pat, rna, gene, 10)
sox10.deg <- rnadeg(sox10.pat, rna)
rnaheat(sox10.pat, rna, sox10.deg, gene)

}

