% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnareact.R
\name{rnareact}
\alias{rnareact}
\title{rnareact function}
\usage{
rnareact(deg)
}
\arguments{
\item{rnadeg}{data frame generated from rnadeg}
}
\value{
data frame of enrichment for Reactome gene sets
}
\description{
Uses ReactomePA to find enriched pathways from the RNAseq differential gene analysis
}
\examples{
#using data from the cureatedTCGAdata set
library(curatedTCGAData)
library(TCGAutils)
lusc <- curatedTCGAData("LUSC", c("RNASeq2GeneNorm"), FALSE)

#split tumor and normal samples
lusc_tn <- splitAssays(lusc, c("01", "11"))
#remake MultiAssayExperiment with only primary tumor samples
lusc_t <- lusc_tn[, , grep("^01", names(lusc_tn))]
lusc_t.egfr <- rnasubset(lusc_t, "EGFR", 10)
egfr_deg <- diffrna(lusc_t.egfr)
egfr_react <- rnareact(egfr_deg)

}
