% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnareact.R
\name{rnareact}
\alias{rnareact}
\title{rnareact function}
\usage{
rnareact(deg, lookup)
}
\arguments{
\item{lookup}{data frame of entrez ID and gene name pairs}

\item{rnadeg}{data frame generated by topTable of limma output}
}
\value{
data frame of enrichment for Reactome gene sets
}
\description{
Uses ReactomePA to find enriched pathways from the RNAseq differential gene analysis
}
\examples{
gene <- "SOX10"
sox10.pat <- patsubset(pat, rna, gene, 10)
sox10.deg <- rnadeg(sox10.pat, rna)
sox10.react <- rnareact(sox10.deg, lookup)

}

