% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patsubset.R, R/rnasubset.R
\name{rnasubset}
\alias{rnasubset}
\alias{rnasubset}
\title{rnasubset function}
\usage{
rnasubset(pat, rna, gene, percent)

rnasubset(pat, rna, gene, percent)
}
\arguments{
\item{pat}{data frame with patient data, each patient is one row}

\item{rna}{data frame with RNAseq count values, genes in rows, patients in columns}

\item{gene}{character(1) Gene symbol}

\item{percent}{numeric(1) percentile of patients to compare 1-50}

\item{pat}{data frame with patient data, each patient is one row}

\item{rna}{data frame with RNAseq count values, genes in rows, patients in columns}

\item{gene}{character(1) Gene symbol}

\item{percent}{numeric(1) percentile of patients to compare 1-50}
}
\value{
Returns a data frame with additional columns for gene of interest expression levels and identifier for high, middle and low groups

Returns a data frame with additional columns for gene of interest expression levels and identifier for high, middle and low groups
}
\description{
Adds gene expression column and gene level columns to patient table

Adds gene expression column and gene level columns to patient table
}
\examples{
data(skcm)
rnasubset(pat, rna, "SOX10", 10)

data(skcm)
rnasubset(pat, rna, "SOX10", 10)

}
