% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppadeg.R
\name{rppadeg}
\alias{rppadeg}
\title{rppadeg function}
\usage{
rppadeg(pat2, rppa)
}
\arguments{
\item{pat2}{data frame generated by rnasubset function}

\item{rppa}{data frame of RPPA values}
}
\value{
data frame of significant differentially expressed proteins between the two groups defined in rnasubset or mutsubset
}
\description{
Uses limma to find differentially expressed proteins between the high and low groups
}
\examples{
data(skcm)
gene <- "SOX10"
sox10.pat <- rnasubset(pat, rna, gene, 10)
sox10.deg <- rppadeg(sox10.pat, rppa)

}
